Renamed widgets/base.py to widgets/widget.py, and updates imports
