+++ src/termui/render_geometry.py
@@ -1,7 +1,7 @@
 from dataclasses import dataclass
 
 from termui.geometry import Region
-from termui.widgets.base import Widget
+from termui.widgets.widget import Widget
 
 
 @dataclass
